Remove difficult to maintain utility_functions. Fixes #1.
The only adaptions to integrate the Convert attack were subtle anyway:

if (type === "Convert") return translate("Convert Attack");

If the convert attack + its stats shall be shown in the GUI, then we
will have to maintain this function, i.e. override it and always
synchronize it with changes in the 0AD equivalent.

This is not necessary for now.